// src/scripts/types/video.type.ts
import { POI } from "./poi.type";

/**
 * An Amadeus Video object
 */
export interface Video {
  /** The id generated by the backend for the video */
  id?: string;
  /** The external id assoicated with this video */
  externalid?: string;
  /** The title of the video */
  title?: string;
  /** The external source for the video */
  source?: string;
  /** The external source for the video */
  original_source?: string;
  /** A list of tags describing the video */
  tags?: string[];
  /** A list of POI appearing in this video */
  video_pois?: POI[];
  /** A list of POI appearing in this video ordered by start time */
  linearized_video_pois?: POI[];
  /** The duration for this video */
  duration?: number;
  /** The user id of this account owning the video */
  producer_id?: string;
  /** A description for this video */
  description?: string;
  /** The date when the video was shot */
  date?: string;
  /** A link to use for the sahre functionality */
  share_link?: string;
  /** An array of caption for the video */
  captions?: any[];
  /** A map of configuration object where each key is a place_id (cf POI) */
  configurations: { [key: string]: any };
  /** A map of pictures where each key is a place_id (cf POI) */
  pictures: { [key: string]: any };
  /** Is the video allowed to be shown */
  enabled?:boolean
}